+++ tests/foreign-crate-registry/lib.rs
@@ -1,0 +1,7 @@
+use stringleton::sym;
+
+stringleton::enable!();
+
+pub fn foo() -> stringleton::Symbol {
+    sym!(foo)
+}